+++ stagewhisper/app/electron/index.ts
@@ -6,7 +6,7 @@ import { join } from 'path';
 import installExtension, { REDUX_DEVTOOLS } from 'electron-devtools-installer';
 
 // Packages
-import { app, BrowserWindow, dialog, ipcMain, IpcMainEvent, IpcMainInvokeEvent } from 'electron';
+import { app, BrowserWindow, dialog, ipcMain, IpcMainEvent } from 'electron';
 import isDev from 'electron-is-dev';
 
 import { existsSync, readFile } from 'fs';
@@ -81,9 +81,10 @@ function createWindow() {
 }
 
 import './handlers/loadVttFromFile';
+import './handlers/runWhisper';
 import './whisperTypes';
-import { spawn } from 'child_process';
-import { WhisperArgs } from './whisperTypes';
+// import { spawn } from 'child_process';
+// import { WhisperArgs } from './whisperTypes';
 
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
@@ -125,27 +126,4 @@ ipcMain.handle('open-directory-dialog', async () => {
   });
 
   return directory.canceled ? null : directory.filePaths[0];
-});
-
-ipcMain.handle('run-whisper', async (_event: IpcMainInvokeEvent, args: WhisperArgs) => {
-  const { inputPath, output_dir } = args;
-  console.log('Running whisper script');
-  console.log('args: ', args);
-
-  // const out = spawn('whisper', ['--model', 'base.en', '--output_dir', join(__dirname, '../src/debug/data')]);
-  const out = spawn('whisper', [inputPath, '--model', 'base.en', '--output_dir', output_dir]);
-
-  out.stdout.on('data', (data) => {
-    console.log(`stdout: ${data}`);
-  });
-  out.stderr.on('data', (err) => {
-    console.log(`stderr: ${err}`);
-  });
-  out.on('message', (message) => {
-    console.log(`message: ${message}`);
-  });
-
-  out.on('close', (code) => {
-    console.log(`child process exited with code ${code}`);
-  });
 });
